refactor: Home screen updated
