+++ babel.config.js
@@ -1,13 +1,13 @@
 module.exports = {
-  presets: ["module:metro-react-native-babel-preset"],
+  presets: ['module:metro-react-native-babel-preset'],
   plugins: [
-    ["react-native-reanimated/plugin"],
+    ['react-native-reanimated/plugin'],
     [
-      "module:react-native-dotenv",
+      'module:react-native-dotenv',
       {
-        envName: "APP_ENV",
-        moduleName: "@env",
-        path: ".env",
+        envName: 'APP_ENV',
+        moduleName: '@env',
+        path: '.env',
         safe: false,
         allowUndefined: true,
         verbose: false,
