+++ src/screens/Home/Home.style.ts
@@ -6,9 +6,4 @@ export const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
   },
-  square: {
-    width: 100,
-    height: 100,
-    backgroundColor: 'blue',
-  },
 });
