+++ src/screens/Home/Home.tsx
@@ -1,78 +1,17 @@
-import { API_URL } from '@env';
-import { NavigationProp, useNavigation } from '@react-navigation/native';
-import React, { useEffect } from 'react';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { FlatList, Text, View } from 'react-native';
-import Animated, {
-  Easing,
-  useAnimatedStyle,
-  useSharedValue,
-  withRepeat,
-  withSpring,
-  withTiming,
-} from 'react-native-reanimated';
-import { BaseLayout, Button, GroupListTile, Header } from '../../components';
-import { IndicatorView } from '../../components/Indicator/Indicator';
+import { FlatList } from 'react-native';
+import { BaseLayout, GroupListTile, Header } from '../../components';
 import { expense } from '../../mocks';
-import { loader } from '../../utils/helper';
-import { styles } from './Home.style';
-import { NavigationParams, THome } from './types';
+import { THome } from './types';
 
 const Home: React.FC<THome> = () => {
-  const navigation = useNavigation<NavigationProp<NavigationParams>>();
   const { t } = useTranslation();
-  const width = useSharedValue(100);
-
-  const handlePress = () => {
-    width.value = withSpring(width.value + 50);
-  };
-
-  const rotation = useSharedValue(0);
-
-  useEffect(() => {
-    const rotateAnimation = () => {
-      rotation.value = withRepeat(
-        withTiming(360, {
-          duration: 2000,
-          easing: Easing.linear,
-        }),
-        -1,
-      );
-    };
-
-    rotateAnimation();
-
-    return () => {
-      rotation.value = 0;
-    };
-  }, [rotation]);
-
-  const animatedStyle = useAnimatedStyle(() => {
-    return {
-      transform: [{ rotate: `${rotation.value}deg` }],
-    };
-  });
 
   return (
     <BaseLayout>
       <Header title={t('home')} />
       <FlatList data={expense.groups} renderItem={({ item }) => <GroupListTile node={item} />} />
-      <Text>
-        {t('env')} {'->'} {API_URL}
-      </Text>
-      <Button title={t('detailsNavigation')} onPress={() => navigation.navigate('Details')} />
-      <IndicatorView isLoading={true} ref={loader} />
-      <Animated.View
-        style={{
-          width,
-          height: 100,
-          backgroundColor: 'violet',
-        }}
-      />
-      <Button onPress={handlePress} title={t('clickMe')} />
-      <View style={styles.container}>
-        <Animated.View style={[styles.square, animatedStyle]} />
-      </View>
     </BaseLayout>
   );
 };
